Overwrite Sender Name and Sender Email using current user's information (Closes #864).
If "Require user registration for replying to an Ad?" setting is enabled, the Reply to Ad
form will pre fill the Sender Name and Sender Email using the current user's profile
information and will disabled the fields.

Also, from now on, the form will show current user's information if available, but will be
editable if the setting mentioned above is disabled.