Restore handler for `wpseo_opengraph_image` filter (Closes #906).
I removed the hook as a solution for #793, but I don't remember the details of the technical problems.
It looks like in that particular version of WordPress SEO, the hook was not being used, or I failed to
see it.

I checked https://github.com/Yoast/wordpress-seo/ and the hook is present in all versions I reviewed. I
cannot explain my decision on #793.

Today's update should work both when the hook is present, and when is not.

The fact that we do not use the `wpseo_opengraph_image` is not really a problem, until someone
defines a default image for the OpenGraph meta tag. At that point, WordPress SEO starts generating
an og:image tag that always overwrites ours, because it is printed first. This update makes sure the
default image is ignored on the Single Listing page, so we are always in control of the image og:image tag.