Pass a configuration array to Media Manager Component.
The idea is to provide information about the different groups of files that are allowed,
so that the JavaScript component doesn't know much about each specific group.

The JavaScript code still knows about videos, images and other files, but the configuration
array (that becomes an options object), will allow better decoupling in the future.