Render all form steps from the beginning.
If a transaction object is available, the Upload Files step
will be shown only if the user is allowed to upload files
for the selected payment term. If no transaction has been
created yet, the step will show up if the plugin settings
allow users to upload images when payment is disabled and
the administrator has marked at least one of the supported
extensions as allowed.