Add helper method to Module class to avoid Fatal error.
Released versions of some modules define module_setup() as a protected method.
We now need that method to be public to run it on init using add_action(), but
changing the access level in this class causes Fatal errors if those modules
are activated. That method is just a workaround.