Fix function that generated media thumbnails URLs (#1174).
The function that returns the thumbnail URL checks whether the file exists or not, but it was trying to do that using the candidate URL as input value, so the function had to replace the site's URL with the absolute path, in order to convert the candidate URL into an absolute path to the thumbnail in the server.

The problem is that in this case, the candidate URLs were being generated with HTTPS (because the site is configure to force secure connection for admin dashboard requests) and the site's URL I was using as a search value was using HTTP (because that was the protocol of the current request).

As a result, the replace operation failed, the URL was not converted to a path, and the file_exists call was returning false. The plugin was unable to find a valid thumbnail URL and returned nothing.

The problem didn't occur to logged in users, or after a user had logged in, because as soon as he or she initiated session, WordPress redirect them to a secure URL and from that moment on, they continue to access all pages using secure URLs.