Add function to render hidden input fields (#1186).
The website was crashing trying to render a WP_Post object as the value
of one of the hidden input fields in the pagination form. I fixed it
excluding all hidden parameters that have objects as values.

The awpcp_html_hidden_fields should be used to render hidden input fields
everywhere.