Add setting to control whether infromation in contact fields is overwritten.
If the setting is enabled, when an administrator is editing a listing and
he changes the selected value in the User/Owner field, the information in
the contact fields (Contact Name, Contact Email and Contact Phone Number)
will be updated (overwriting the information already entered in those fields)
using the infromation of the user just selected. The modifications will not
be persisted until the Save Changes button is clicked.