Move code to handle setregion requests to Region Control module (#1213).
The setregion request was previously handled in the code that renders the output for
the main classifieds pages. However, at that point it was not possible to update the
active region, because we now store that information in a cookie, and cookies need to
be sent before the page's output is sent.