Update dynamic breakpoints for Media Manager.
If the available space is wider than 992px we use a fixed size for each image,
to allow the browser to put as many images in a single row as possible. It
looks better, IMO, even though they are not longer aligned with the invisible
borders of the container DIV.

I also had to show fewer elements per row in smaller screens, to accommodate
the new toggle button used to approve and reject media files.