Allow ZIP files that extract into a directory with the name of the ZIP file (#1383).
Previously, the only way to use those kind of files was to enter the full path to the image
in the CSV file (my-zip-file/my-photo.jpg if the photo was part of my-zip-file.zip).

Now you can enter my-photo.jpg and the importer will handle the rest.

Note that if the name of the extracted directory is different than the name of the ZIP file,
then you still have to enter the full path to the image in the CSV file.