Fix Random Listings shortcode (#1542).
The shortcode was picking a random listings from the last N listings,
where N was the deafult number of listings to show in the Listings page.

Listings Finder is designed to use the deafult number of listings to show in
the Listings page when the query 'limit' is set to 0 (the default value).

Setting the limit to false causes the query not to include a LIMIT clause.