Fix error in code that renders content placeholders.
The placeholders have to be sorted in reverse alphabetical order,
to prevent a the content of a placeholder such as `$ratings` from
being rendered as part of the output of `$ratings_readonly`.

Since both placeholders start with the same string, if both of
them are being used and you try to replace the shorter one with
generated content, the `$ratings` part of the longer one will
also match and be replaced with that content; causing the longer
placeholder to break.