Rename 'awpcp_category' parameter to 'awpcp_category_id'. (#2293)
Fixes the behavior of the Category filter in Classifieds Ads admin
screen, that got broken after 383bb66a4 was introduced in #2277.

'awpcp_category' is now the query var used to request terms in
the the Listing Category taxonomy. It shouldn't be used as a
regular parameter. If used, it should be set to the slug of an
existing category instead of an ID.